homework-1/pyproject.toml
[tool.poetry]
name = "homework-1"
version = "0.1.0"
description = ""
authors = ["Saysuu <skysaika@yandex.ru>"]
readme = "README.md"
packages = [{include = "homework_1"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
